+++ R/as.data.frame.R
@@ -1,8 +1,11 @@
-#' Coerce a surveytable table to a data frame
+#' Coerce an astra table to a data frame
 #'
 #' @description
 #'
-#' If a tabulation function produces multiple tables, that group of tables is a list,
+#' Coerce an astra table to a data frame. To restructure tables to make them
+#' easier to process programmatically, see `restructure()`. Also see `set_opts(raw = TRUE)`.
+#'
+#' If a tabulation function produces multiple tables, that group of tables is a `list`,
 #' with each element of the list being an individual table. To convert one of these tables
 #' to a `data.frame`, use `[[`. For example, in the following code, we generate
 #' 3 tables, and then convert the third table to a `data.frame`.
@@ -18,12 +21,14 @@
 #' @returns
 #' A data frame.
 #'
+#' @family print
+#'
 #' @export
 #'
 #' @examples
 #' set_survey(namcs2019sv)
 #' as.data.frame( tab("AGER") )
-as.data.frame.surveytable_table = function(x, ...) {
+as.data.frame.astra_table = function(x, ...) {
   class(x) = "data.frame"
   names(x) = make.names( names(x), unique = TRUE )
   x