+++ R/z_utils.R
@@ -31,21 +31,6 @@
   table[idx]
 }
 
-# huxtable:::assert_package
-assert_package = function (fun, package, version = NULL)
-{
-  if (!requireNamespace(package, quietly = TRUE))
-    stop(glue::glue("`{fun}` requires the \"{package}\" package. To install, type:\n",
-                    "install.packages(\"{package}\")"))
-  if (!is.null(version)) {
-    cur_ver <- utils::packageVersion(package)
-    if (cur_ver < version)
-      stop(glue::glue("`{fun}` requires version {version} or higher of the \"{package}\" ",
-                      "package. You have version {cur_ver} installed. To update the package,",
-                      "type:\n", "install.packages(\"{package}\")"))
-  }
-}
-
 o2s = function(obj) {
   glue_collapse(class(obj), sep = ", ")
 }
@@ -71,15 +56,18 @@ o2s = function(obj) {
 # show_options() |> names() |> dput()
 #### !!! If making changes, update: .onLoad(), set_opts(), show_opts(), .check_options()
 .check_options = function() {
-  c_need = c("surveytable.drop_na", "surveytable.find_lpe",
+  c_need = c("astra.file", "astra.file_show", "astra.print",
+             "surveytable.drop_na", "surveytable.find_lpe",
              "surveytable.lpe_counts", "surveytable.lpe_n", "surveytable.lpe_percents",
              "surveytable.max_levels", "surveytable.names_count", "surveytable.names_count_raw",
-             "surveytable.names_prct", "surveytable.print", "surveytable.file", "surveytable.file_show",
+             "surveytable.names_prct",
              "surveytable.p.adjust_method", "surveytable.rate_per", "surveytable.raw",
              "surveytable.survey_label", "surveytable.svychisq_statistic",
              "surveytable.svyciprop_adj", "surveytable.tx_count", "surveytable.tx_df",
              "surveytable.tx_numeric", "surveytable.tx_prct", "surveytable.tx_pval",
-             "surveytable.tx_rate", "surveytable.tx_test_stat")
+             "surveytable.tx_rate", "surveytable.tx_test_stat",
+             "surveytable.restructure_attr", "surveytable.restructure_flags"
+             , "surveytable.age_adjusted")
   c_have = show_options() %>% names()
 
   d_nh = setdiff(c_need, c_have)